# Optimize some operations in points.py, fix == for non-points