This optimizes some operations in points.py significantly by using functions from Python's math library instead of numpy (numpy is much faster for arrays, but in points.py, all use-cases are on scalar values, so we can use Python's math library instead) - we could replace sqrt, cos, sin, arccos, radians, degrees and isnan, which helped acheive significant performance boosts in several areas.

Additionally, using exponentiation to square numbers is quite slow - a multiplication is much faster, and this helps makes speed things up as well.

This also fixes equality checks (so that they don't error on non-Points), and also makes it faster by using np.allclose over isclose.

The timings for various operations, over 10000 iterations, are as follows (in seconds)

```
New == 0.302957s
Old == 0.803580s
New length 0.005106s
Old length 0.021221s
New angle_opp 0.004432s
Old angle_opp 0.031315s
New angle_between 0.014184s
Old angle_between 0.047519s
New project_vec 0.020047s
Old project_vec 0.023098s
New _return_sin_cos 0.003705s
Old _return_sin_cos 0.030704s
```